refactor: 사용자 쿼리 관련 파일 구조 변경
